add bandwidths to graph
Add bandwidths to graph to enable multi-path routing in the future.
Since Ryu can't store auxiliary link data such as bandwidth, the
topology learning code cannot learn the bandwidths of the links. Now
there is a file bandwidths.edgelist that contains an edgelist of the
graph. The topology discovery code reads the edgelist and adds the
bandwidth information to known existing links.

This commit also removes the obsolete 'latency' field from the topology
file since we decided hopCount to be a better metric than latency for
numerous reasons.