include bw in graph, calculate mpls labels
integrate multipath_labelSwap into the controller.
graph now properly contains bandwidths from 'bandwidths.edgelist'

note: updating the topology will invalidate existing labels, so the
graph must be rebuilt and the routes/labels recalculated on a topology
update. this also means that existing flows based on the labels (not yet
implemented) must be deleted; a broken function for deleting existing
flows is implemented but unused in this commit.