fix ARP handling
popping the MPLS header does not set the packet's ethertype to whatever
it was before the header was pushed -- it sets it to IP by default.
this means that ARP replies, which are unicast, were being incorrectly
set as IP packets when popping the MPLS header, causing the reply to
fail at the end host.

to handle this, the label-swapping code now generates an additional path
to each destination that is identical to the path with the lowest hop
count except with a different label to indicate it is the ARP path.

ARP path labels have values in the range [100, 1000) and non-ARP path
labels have values of at least 1000 so the controller can differentiate
between ARP and IP traffic in order to assign the correct ethertype when
popping MPLS headers.