fix topology bug
when topology is cleared and rebuilt, pings flood forever and flows are
not added to the switches. this is because self.mac_to_port was not
being cleared so no switches were running the code to add flows for
forwarding packets to their connected hosts. this commit clears
mac_to_port during topology updates and fixes this bug.