add known hosts flow table
In order for hosts to create the source-based flood tree for each host,
they have to know about every host. This commit ensure switches will
know about every host that sends traffic through the switch.

This commit adds a "known hosts" flow table. It matches on the eth src
field and the only instruction is to go to the primary routing table. If
a match is not found, the packet is sent to the controller, which will
then install an entry in the known hosts table for this switch and
install entries in the routing table for flooding/dropping broadcast
traffic depending on the port.